Adding better nav, signin flow

+++ app/controllers/user_sessions_controller.rb
@@ -1,6 +1,5 @@
 class UserSessionsController < ApplicationController
   before_filter :require_no_user, :only => [:new, :create]
-  before_filter :require_user, :only => :destroy
   
   def new
     @user_session = UserSession.new
@@ -17,7 +16,7 @@ class UserSessionsController < ApplicationController
   end
   
   def destroy
-    current_user_session.destroy
+    current_user_session.destroy if current_user_session
     flash[:notice] = "Logout successful!"
     redirect_back_or_default new_user_session_url
   end
